tests: add FullIonParquetConversionTest
